add omitempty option
This change adds a `gogoproto.omitempty` option that can be used (in
conjunction with `(gogoproto.nullable) = false`) for non-nullable
message fields.

When a message field is not nullable, its tag is always encoded - even
if all message fields are unset. The `omitempty` option changes the
marshalling code to check an `Empty()` method on that message, which
must be defined by the user. The result is that when the message is
empty, the encoding matches the one we would get if the field was
nullable and the pointer is unset.