unmarshal: reuse packed slices when possible
When unmarshaling a packed slice, we reuse the slice if possible
(instead of always allocating a new one). This change is a no-op in
most cases where we use `proto.Unmarshal` which calls `Reset()` in the
message. However, it allows optimized paths to use the `Unmarshal`
method directly and avoid allocations.