omitempty: use IsEmpty instead of Empty
The omitempty option will be used to clean up the
MVCCValueHeaderPure/MVCCValueHeaderCrdbTest hack.

This commit changes the generated code to rely on an `IsEmpty()`
method (instead of `Empty`). Timestamp already has an `IsEmpty()`
method which is used in many places; plus "is empty" is better
("empty" can be interpreted as an imperative verb).