oneof: batch allocate oneof wrapper and field messages
This commit updates the unmarshalling logic for oneof messages to avoid a pair
of heap allocations. Instead, we allocate a single object that contains both the
oneof wrapper and the field message. We then point the wrapper to the field in
the same heap object.

We were already playing this trick in CockroachDB. We'll benefit even more from
it by adding it to gogoproto generated code.